// frontend/src/modules/tasks/types.ts
export interface User {
  id: string;
  name: string;
  email: string;
  avatar?: string;
}

export interface Task {
  id: string;
  title: string;
  description: string;
  dueDate: string;
  priority: 'LOW' | 'MEDIUM' | 'HIGH' | 'URGENT';
  status: 'TODO' | 'IN_PROGRESS' | 'REVIEW' | 'COMPLETED';
  creatorId: string;
  assignedToId?: string;
  createdAt: string;
  updatedAt: string;
  creator: User;
  assignedTo?: null | User ;
  createdById: string; 
  
}

export interface CreateTaskData {
  title: string;
  description?: string;
  dueDate: string;
  priority: Task['priority'];
  status: Task['status'];
  assignedToId?: string | null;
}

export interface UpdateTaskData extends Partial<CreateTaskData> {}